deconstruct onExpand
